ts-web: move NaclSigner to signer-tweetnacl
