+++ client-sdk/ts-web/core/src/signature.ts
@@ -1,5 +1,3 @@
-import * as nacl from 'tweetnacl';
-
 import * as hash from './hash';
 import * as misc from './misc';
 import * as types from './types';
@@ -111,68 +109,6 @@ export class BlindContextSigner implements ContextSigner {
     async sign(context: string, message: Uint8Array): Promise<Uint8Array> {
         const signerMessage = prepareSignerMessage(context, message);
         return await this.signer.sign(signerMessage);
-    }
-}
-
-/**
- * An in-memory signer based on tweetnacl. We've included this for development.
- */
-export class NaclSigner implements Signer {
-    key: nacl.SignKeyPair;
-
-    constructor(key: nacl.SignKeyPair, note: string) {
-        if (note !== 'this key is not important') throw new Error('insecure signer implementation');
-        this.key = key;
-    }
-
-    /**
-     * Generate a keypair from a random seed
-     * @param note Set to 'this key is not important' to acknowledge the risks
-     * @returns Instance of NaclSigner
-     */
-    static fromRandom(note: string) {
-        const secret = new Uint8Array(32);
-        crypto.getRandomValues(secret);
-        return NaclSigner.fromSeed(secret, note);
-    }
-
-    /**
-     * Instanciate from a given secret
-     * @param secret 64 bytes ed25519 secret (h) that will be used to sign messages
-     * @param note Set to 'this key is not important' to acknowledge the risks
-     * @returns Instance of NaclSigner
-     */
-    static fromSecret(secret: Uint8Array, note: string) {
-        const key = nacl.sign.keyPair.fromSecretKey(secret);
-        return new NaclSigner(key, note);
-    }
-
-    /**
-     * Instanciate from a given seed
-     * @param seed 32 bytes ed25519 seed (k) that will deterministically generate a private key
-     * @param note Set to 'this key is not important' to acknowledge the risks
-     * @returns Instance of NaclSigner
-     */
-    static fromSeed(seed: Uint8Array, note: string) {
-        const key = nacl.sign.keyPair.fromSeed(seed);
-        return new NaclSigner(key, note);
-    }
-
-    /**
-     * Returns the 32 bytes public key of this key pair
-     * @returns Public key
-     */
-    public(): Uint8Array {
-        return this.key.publicKey;
-    }
-
-    /**
-     * Signs the given message
-     * @param message Bytes to sign
-     * @returns Signed message
-     */
-    async sign(message: Uint8Array): Promise<Uint8Array> {
-        return nacl.sign.detached(message, this.key.secretKey);
     }
 }
